pyproject.toml
[tool.poetry]
name = "alkemio-virtual-contributor"
version = "0.1.0"
description = "Alkemio Virtual Contributor — unified microkernel engine with pluggable handlers"
authors = ["Alkemio BV <info@alkem.io>"]
license = "EUPL-1.2"
readme = "README.md"
packages = [{include = "core"}, {include = "plugins"}]

[tool.poetry.dependencies]
python = "^3.12"

# Messaging
aio-pika = "9.5.7"
aiormq = "6.9.0"
async-timeout = "5.0.1"

# Configuration
pydantic = "^2.11"
pydantic-settings = "^2.11.0"
load-dotenv = "^0.1.0"

# LLM / AI
langchain = "^1.1.0"
langchain-openai = "^1.1.0"
langchain-mistralai = "^1.1.0"
langgraph = "^1.0.4"
langchain-text-splitters = "^0.3.8"
openai = "^1.109"

# Vector DB
chromadb-client = "^1.5.0"

# Utilities
json-schema-to-pydantic = "^0.4.3"
bcrypt = "^5.0.0"
httpx = "^0.27.2"
beautifulsoup4 = "^4.14"
requests = "^2.32"

# File parsers (ingest-space)
pypdf = "^5.0"
python-docx = "^1.1"
openpyxl = "^3.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pyright = "^1.1.406"
pytest = "^9.0"
pytest-cov = "^7.1"
pytest-asyncio = "^1.3"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
omit = [
    "core/adapters/*",
    "core/config.py",
    "core/health.py",
    "core/logging.py",
    "main.py",
    "plugins/ingest_space/graphql_client.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
